Fixed a couple of problems with the storage's open() method.
Previously, open() wasn't accepting *args (only **kwargs), which caused
crashes when called as foo.open(filename, "r"), for example.

Also, it was only attempting to open the locally saved copy, which may
not even exist for something that has already been sent to remote
storage. So now we try remote first and then local, the same as for all
the other methods.